+++ src/pipecatcloud/cli/commands/agent.py
@@ -3,15 +3,17 @@ import questionary
 import typer
 from loguru import logger
 from rich import box
-from rich.console import Console, Group
+from rich.console import Group
+from rich.columns import Columns
 from rich.live import Live
 from rich.panel import Panel
 from rich.table import Table
+from rich.text import Text
 
-from pipecatcloud._utils.agent_utils import handle_agent_start_error
+from enum import Enum
 from pipecatcloud._utils.async_utils import synchronizer
 from pipecatcloud._utils.auth_utils import requires_login
-from pipecatcloud._utils.console_utils import console
+from pipecatcloud._utils.console_utils import console, format_timestamp
 from pipecatcloud.cli import PIPECAT_CLI_NAME
 from pipecatcloud.cli.api import API
 from pipecatcloud.cli.config import config
@@ -104,29 +106,6 @@ async def status(
             console.error(f"No deployment data found for agent with name '{agent_name}'")
             return typer.Exit()
 
-        # Conditions
-        conditions = data["conditions"]
-        table = Table(
-            show_header=True,
-            show_lines=True,
-            border_style="dim",
-            box=box.SIMPLE
-        )
-        table.add_column("Status")
-        table.add_column("Type")
-        table.add_column("Message")
-        table.add_column("Reason")
-        table.add_column("Date")
-
-        for condition in conditions:
-            table.add_row(
-                DEPLOY_STATUS_MAP.get(condition["status"], "[dim]Unknown[/dim]"),
-                condition['type'],
-                condition.get('message', 'No message'),
-                condition.get('reason', 'No reason'),
-                condition['lastTransitionTime']
-            )
-
         # Deployment info
 
         deployment_table = Table(
@@ -137,12 +116,12 @@ async def status(
         deployment_table.add_column("Key")
         deployment_table.add_column("Value")
         deployment_table.add_row(
-            "[bold]Image:[/bold]",
-            str(data.get("deployment", {}).get("manifest", {}).get("spec", {}).get("image", "N/A")),
-        )
-        deployment_table.add_row(
             "[bold]Active Session Count:[/bold]",
             str(data.get("activeSessionCount", "N/A")),
+        )
+        deployment_table.add_row(
+            "[bold]Image:[/bold]",
+            str(data.get("deployment", {}).get("manifest", {}).get("spec", {}).get("image", "N/A")),
         )
         deployment_table.add_row(
             "[bold]Active Deployment ID:[/bold]",
@@ -160,24 +139,17 @@ async def status(
         # Autoscaling info
         autoscaling_data = data.get("autoScaling", None)
         if autoscaling_data:
-            autoscaling_table = Table(
-                show_header=False,
-                show_lines=False,
-                box=box.SIMPLE
-            )
-            autoscaling_table.add_column("Key")
-            autoscaling_table.add_column("Value")
-
-            autoscaling_table.add_row(
-                "[bold]Max instances:[/bold]",
-                str(autoscaling_data.get("maxReplicas", 0)),
-            )
-            autoscaling_table.add_row(
-                "[bold]Min instances:[/bold]",
-                str(autoscaling_data.get("minReplicas", 0)),
-            )
-            autoscaling_panel = Panel(
-                autoscaling_table,
+            scaling_renderables = [
+                Panel(
+                    f"[bold]Minimum Instances[/bold]\n{autoscaling_data.get('minReplicas', 0)}",
+                    expand=True),
+                Panel(
+                    f"[bold]Maximum Instances[/bold]\n{autoscaling_data.get('maxReplicas', 0)}",
+                    expand=True)]
+            scaling_panel = Panel(
+                Columns(
+                    scaling_renderables
+                ),
                 title="[bold]Scaling configuration:[/bold]",
                 title_align="left",
                 border_style="dim",
@@ -190,8 +162,7 @@ async def status(
             Panel(
                 Group(
                     deployment_table,
-                    autoscaling_panel if autoscaling_panel else "",
-                    table,
+                    scaling_panel if scaling_panel else "",
                     Panel(
                         f"[{color}]Health: {'Ready' if data['ready'] else 'Stopped'}[/]",
                         border_style="green" if data['ready'] else "yellow",
@@ -252,31 +223,69 @@ async def scale():
     console.error("Not implemented")
 
 
+class LogLevel(str, Enum):
+    DEBUG = "DEBUG"
+    INFO = "INFO"
+    WARNING = "WARNING"
+    ERROR = "ERROR"
+    CRITICAL = "CRITICAL"
+
+
+class LogLevelColors(str, Enum):
+    DEBUG = "dim white"
+    INFO = "white"
+    WARNING = "yellow"
+    ERROR = "red"
+    CRITICAL = "bold red"
+
+
 @agent_cli.command(name="logs", help="Get logs for the given agent.")
 @synchronizer.create_blocking
 @requires_login
-async def logs(ctx: typer.Context, agent_name: str, organization: str = typer.Option(
-    None,
-    "--organization",
-    "-o",
-    help="Organization to get status of agent for"
-),):
+async def logs(
+    agent_name: str,
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o",
+        help="Organization to get status of agent for"
+    ),
+    level: LogLevel = typer.Option(
+        None,
+        "--level", "-l",
+        help="Level of logs to get"
+    ),
+    limit: int = typer.Option(
+        100,
+        "--limit", "-n",
+        help="Number of logs to get"
+    ),
+):
     org = organization or config.get("org")
-    token = config.get("token")
 
-    try:
-        with console.status(f"Fetching logs for agent: [bold]'{agent_name}'[/bold]", spinner="dots"):
-            async with aiohttp.ClientSession() as session:
-                response = await session.get(
-                    f"{API.construct_api_url('services_logs_path').format(org=org, service=agent_name)}?limit=100&order=desc",
-                    headers={"Authorization": f"Bearer {token}"},
-                )
-                if response.status != 200:
-                    response.raise_for_status()
-                data = await response.json()
-                console.print(data)
-    except Exception:
-        console.error(f"Unable to get logs for {agent_name}")
+    with console.status(f"[dim]Fetching logs for agent: [bold]'{agent_name}'[/bold] with severity: [bold cyan]{level.value if level else 'ALL'}[/bold cyan][/dim]", spinner="dots"):
+        data, error = await API.agent_logs(agent_name=agent_name, org=org, limit=limit)
+
+        if not data or not data.get("logs"):
+            console.print("[dim]No logs found for agent[/dim]")
+            return typer.Exit(1)
+
+    for l in data["logs"]:
+        log_data = l.get("log", "")
+        if log_data:
+            timestamp = format_timestamp(l.get("timestamp", ""))
+            severity = LogLevel.INFO
+            for log_severity in LogLevel:
+                if log_severity.value in log_data.upper():
+                    severity = log_severity
+                    break
+            # filter out any messages that do not match our log level
+            if level and severity.value != level.value:
+                continue
+            color = getattr(LogLevelColors, severity, LogLevelColors.DEBUG).value
+            console.print(Text(timestamp, style="bold dim"))
+            console.print(Text(l.get("log", ""), style=color))
+            console.rule(style="dim")
 
 
 @agent_cli.command(name="delete", help="Delete an agent.")
@@ -486,7 +495,7 @@ async def start(
             daily_room = data.get("dailyRoom")
             daily_token = data.get("dailyToken")
             if daily_room:
-                message += f"\n\nDaily room: [link={daily_room}?t={daily_token}]{daily_room}?t={daily_token}[/link]"
+                message += f"\n\nDaily room: [link={daily_room}?t={daily_token}]{daily_room}?t={daily_token}[/link]\n"
             return console.success(
                 message,
                 subtitle=f"[white dim]Join the session:[/white dim] [link={daily_room}?t={daily_token}]click here[/link]")