+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to **Pipecat Cloud** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.10] - 2025-03-04
+
+### Added
+- `init` convenience command will now populate the working directory with files from the starter project
+- `agent log` allows for optional severity level filtering
+
+### Changed
+- `agent status` and `deploy` no longer show superfluous data
+- `session.py` now correctly handles errors when starting agents
+- `secrets set` no longer prompts twice for confirmation if the secret set does not exist
+
+### Removed
+- `errors.py` removed as redundant (error message and code returned via API)
+- `agent_utils.py` removed as redundant (error message and code returned via API)
+
 ## [0.0.9] - 2025-02-27
 
 ### Added